New Smartbar icons to match Pixel theme
A few weeks ago we made the change system wide from Material teal to
Pixel blue. This changed was welcomed with open arms by many.  With this
change came the pixel navbar icons for 'back/home/recents' and while
this change was also welcomed, it looked off (at least to us) when you
enabled other icons.

So what we did was reach out to the talented Rahul K Dinesh​ and he delivered!

He has made beautiful icons for us that fit perfectly with the traditional
pixel navbar icons and we couldn't be happier :-)

Make sure to check out his other work in the playstore and on Google+!

https://play.google.com/store/apps/details?id=com.b16h22.rubiq
https://play.google.com/store/apps/details?id=com.b16h22.navswag

https://plus.google.com/u/1/106222165086709621824

 Conflicts:
	res/drawable/ic_sysbar_one_handed_mode_left.xml
	res/drawable/ic_sysbar_one_handed_mode_right.xml

Change-Id: I774b750726cdcd437c3cf0c0d5e002d601cf139b